[Refactor] Split morsel queue implementations into focused files
