+++ be/src/exec/pipeline/scan/fixed_morsel_queue.h
@@ -1,0 +1,38 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <atomic>
+
+#include "exec/pipeline/scan/morsel_queue.h"
+
+namespace starrocks::pipeline {
+
+// The morsel queue with a fixed number of morsels, which is determined in the constructor.
+class FixedMorselQueue final : public MorselQueue {
+public:
+    explicit FixedMorselQueue(Morsels&& morsels) : MorselQueue(std::move(morsels)), _pop_index(0) {}
+    ~FixedMorselQueue() override = default;
+    bool empty() const override { return _unget_morsel == nullptr && _pop_index >= _num_morsels; }
+    StatusOr<MorselPtr> try_get() override;
+
+    std::string name() const override { return "fixed_morsel_queue"; }
+    Type type() const override { return FIXED; }
+
+private:
+    std::atomic<size_t> _pop_index;
+};
+
+} // namespace starrocks::pipeline
